+++ test/fixtures/reject.undefined-functions.disallowed-functions.vue
@@ -3,5 +3,6 @@
 <style lang="scss">
 .becky {
 	min-height: rem(10);
+	height: v-bind(height);
 }
 </style>